requirements.txt
pytest
pytest-playwright
playwright
pyyaml
allure-pytest
pytest-xdist
pytest-rerunfailures
